# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "此 API 文件提供社區通知系統的後端介面說明，包含使用者管理、設施預約、社區申請及訊息通知等功能。使用 JWT Bearer 進行認證。",
        "title": "社區通知系統 API 文件 (Community Notification System)",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/admin/reschedule/{id}/approve": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理員針對住戶提出的改期申請進行核准或拒絕。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reservation"
                ],
                "summary": "審核預約改期 (管理員)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "改期申請紀錄 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "審核結果與意見",
                        "name": "approve",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reservation.ApproveRescheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "改期申請處理完成",
                        "schema": {
                            "$ref": "#/definitions/model.RequestMessage"
                        }
                    },
                    "400": {
                        "description": "無效的 ID 或輸入資料",
                        "schema": {
                            "$ref": "#/definitions/model.Response400Error"
                        }
                    },
                    "500": {
                        "description": "審核處理失敗",
                        "schema": {
                            "$ref": "#/definitions/model.Response500Error"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/reservations/{id}/delete": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理員軟刪除同社區預約，住戶列表將不再顯示該預約。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reservation"
                ],
                "summary": "管理員強制取消預約",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "預約 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "預約刪除成功",
                        "schema": {
                            "$ref": "#/definitions/model.RequestMessage"
                        }
                    },
                    "400": {
                        "description": "無效的預約 ID",
                        "schema": {
                            "$ref": "#/definitions/model.Response400Error"
                        }
                    },
                    "401": {
                        "description": "無法取得登入資訊",
                        "schema": {
                            "$ref": "#/definitions/model.Response401Error"
                        }
                    },
                    "403": {
                        "description": "權限不足",
                        "schema": {
                            "$ref": "#/definitions/model.Response403Error"
                        }
                    },
                    "404": {
                        "description": "找不到預約",
                        "schema": {
                            "$ref": "#/definitions/model.Response404Error"
                        }
                    }
                }
            }
        },
...
        "/api/v1/facilities/{facility_id}/reservations": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "住戶針對特定社區設施進行預約。系統會根據設施規則（如黑名單、預約時段重疊等）進行驗證。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reservation"
                ],
                "summary": "建立設施預約",
                "parameters": [
                    {
                        "type": "string",
                        "description": "社區 ID",
                        "name": "community_id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "設施 ID",
                        "name": "facility_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "預約詳細資料",
                        "name": "reservation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reservation.CreateReservationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "預約建立成功",
                        "schema": {
                            "$ref": "#/definitions/model.RequestMessage"
                        }
                    },
                    "400": {
                        "description": "無效的設施 ID 或輸入資料",
                        "schema": {
                            "$ref": "#/definitions/model.Response400Error"
                        }
                    },
                    "403": {
                        "description": "設施不開放或權限不足",
                        "schema": {
                            "$ref": "#/definitions/model.Response403Error"
                        }
                    },
                    "404": {
                        "description": "設施不存在",
                        "schema": {
                            "$ref": "#/definitions/model.Response404Error"
                        }
                    },
                    "409": {
                        "description": "時段衝突或違反設施規則",
                        "schema": {
                            "$ref": "#/definitions/model.Response409Error"
                        }
                    }
                }
            }
        },
...
        "/api/v1/reservations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "依 JWT 社區取得預約列表。住戶只會看到自己的預約；管理員可看到同社區全部未刪除預約。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reservation"
                ],
                "summary": "取得預約列表",
                "responses": {
                    "200": {
                        "description": "預約列表",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "無法取得登入資訊",
                        "schema": {
                            "$ref": "#/definitions/model.Response401Error"
                        }
                    },
                    "500": {
                        "description": "查詢預約列表失敗",
                        "schema": {
                            "$ref": "#/definitions/model.Response500Error"
                        }
                    }
                }
            }
        },
        "/api/v1/reservations/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "依 JWT 社區與角色範圍取得單一預約詳情。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reservation"
                ],
                "summary": "取得預約詳情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "預約 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "預約詳情",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "無效的預約 ID",
                        "schema": {
                            "$ref": "#/definitions/model.Response400Error"
                        }
                    },
                    "401": {
                        "description": "無法取得登入資訊",
                        "schema": {
                            "$ref": "#/definitions/model.Response401Error"
                        }
                    },
                    "404": {
                        "description": "找不到預約",
                        "schema": {
                            "$ref": "#/definitions/model.Response404Error"
                        }
                    }
                }
            }
        },
        "/api/v1/reservations/{id}/cancel": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "住戶主動取消已建立的設施預約。需提供取消原因。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reservation"
                ],
                "summary": "取消預約",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "預約紀錄 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "取消原因",
                        "name": "cancel",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reservation.CancelReservationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "預約已成功取消",
                        "schema": {
                            "$ref": "#/definitions/model.RequestMessage"
                        }
                    },
                    "400": {
                        "description": "無效的 ID 或取消失敗",
                        "schema": {
                            "$ref": "#/definitions/model.Response400Error"
                        }
                    }
                }
            }
        },
        "/api/v1/reservations/{id}/reschedule": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "住戶針對現有的預約申請變更日期或時段。申請後需經由管理員審核。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reservation"
                ],
                "summary": "申請預約改期",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "預約紀錄 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "新的預約時段與原因",
                        "name": "reschedule",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reservation.RescheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "改期申請已送出",
                        "schema": {
                            "$ref": "#/definitions/model.RequestMessage"
                        }
                    },
                    "400": {
                        "description": "無效的 ID 或輸入資料",
                        "schema": {
                            "$ref": "#/definitions/model.Response400Error"
                        }
                    },
                    "404": {
                        "description": "找不到預約紀錄",
                        "schema": {
                            "$ref": "#/definitions/model.Response404Error"
                        }
                    },
                    "409": {
                        "description": "改期衝突或處理失敗",
                        "schema": {
                            "$ref": "#/definitions/model.Response409Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.RequestMessage": {
            "type": "object",
            "properties": {
                "Message": {
                    "type": "string",
                    "example": "message"
                }
            }
        },
        "model.Response400Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "type": "string",
                    "example": "Invalid input data"
                },
                "internal_code": {
                    "type": "integer",
                    "example": 1001
                },
                "request_id": {
                    "type": "string",
                    "example": "uuid-string"
                },
                "status": {
                    "type": "string",
                    "example": "Bad Request"
                }
            }
        },
        "model.Response401Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 401
                },
                "error": {
                    "type": "string",
                    "example": "Invalid or missing JWT token"
                },
                "internal_code": {
                    "type": "integer",
                    "example": 1002
                },
                "request_id": {
                    "type": "string",
                    "example": "uuid-string"
                },
                "status": {
                    "type": "string",
                    "example": "Unauthorized"
                }
            }
        },
        "model.Response403Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 403
                },
                "error": {
                    "type": "string",
                    "example": "Permission denied"
                },
                "internal_code": {
                    "type": "integer",
                    "example": 1003
                },
                "request_id": {
                    "type": "string",
                    "example": "uuid-string"
                },
                "status": {
                    "type": "string",
                    "example": "Forbidden"
                }
            }
        },
        "model.Response404Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 404
                },
                "error": {
                    "type": "string",
                    "example": "Resource not found"
                },
                "internal_code": {
                    "type": "integer",
                    "example": 1004
                },
                "request_id": {
                    "type": "string",
                    "example": "uuid-string"
                },
                "status": {
                    "type": "string",
                    "example": "Not Found"
                }
            }
        },
        "model.Response409Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 409
                },
                "error": {
                    "type": "string",
                    "example": "Resource conflict"
                },
                "internal_code": {
                    "type": "integer",
                    "example": 1005
                },
                "request_id": {
                    "type": "string",
                    "example": "uuid-string"
                },
                "status": {
                    "type": "string",
                    "example": "Conflict"
                }
            }
        },
        "model.Response500Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "error": {
                    "type": "string",
                    "example": "Internal server error"
                },
                "internal_code": {
                    "type": "integer",
                    "example": 1000
                },
                "request_id": {
                    "type": "string",
                    "example": "uuid-string"
                },
                "status": {
                    "type": "string",
                    "example": "Internal Server Error"
                }
            }
        },
...
        "reservation.ApproveRescheduleRequest": {
            "type": "object",
            "properties": {
                "admin_comment": {
                    "type": "string",
                    "example": "已核准改期申請"
                },
                "is_approve": {
                    "description": "true: 核准, false: 拒絕",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "reservation.CancelReservationRequest": {
            "type": "object",
            "properties": {
                "cancel_reason": {
                    "type": "string",
                    "example": "臨時有事無法前往"
                }
            }
        },
        "reservation.CreateReservationRequest": {
            "type": "object",
            "required": [
                "end_time",
                "people_count",
                "reservation_date",
                "start_time"
            ],
            "properties": {
                "end_time": {
                    "description": "HH:mm",
                    "type": "string",
                    "example": "15:00"
                },
                "home_id": {
                    "description": "可選，看業務邏輯",
                    "type": "integer",
                    "example": 1
                },
                "people_count": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 2
                },
                "remark": {
                    "type": "string",
                    "example": "使用跑步機"
                },
                "reservation_date": {
                    "description": "YYYY-MM-DD",
                    "type": "string",
                    "example": "2025-05-20"
                },
                "start_time": {
                    "description": "HH:mm",
                    "type": "string",
                    "example": "14:00"
                }
            }
        },
        "reservation.RescheduleRequest": {
            "type": "object",
            "required": [
                "new_date",
                "new_end_time",
                "new_start_time"
            ],
            "properties": {
                "new_date": {
                    "type": "string",
                    "example": "2025-05-21"
                },
                "new_end_time": {
                    "type": "string",
                    "example": "11:00"
                },
                "new_start_time": {
                    "type": "string",
                    "example": "10:00"
                },
                "reason": {
                    "type": "string",
                    "example": "原本時段突然要開會"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}